remove old connections from the plugin, promote new connections
