+++ acapy_plugin_toolbox/connections.py
@@ -3,14 +3,16 @@
 # pylint: disable=invalid-name
 # pylint: disable=too-few-public-methods
 
-from marshmallow import fields, validate
+from typing import Dict, Any
+
+from marshmallow import Schema, fields, validate
 
 from aries_cloudagent.config.injection_context import InjectionContext
 from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 from aries_cloudagent.messaging.base_handler import BaseHandler, BaseResponder, RequestContext
 from aries_cloudagent.protocols.connections.manager import ConnectionManager
 from aries_cloudagent.connections.models.connection_record import (
-    ConnectionRecord, ConnectionRecordSchema
+    ConnectionRecord
 )
 from aries_cloudagent.protocols.connections.messages.connection_invitation import (
     ConnectionInvitation,
@@ -20,300 +22,160 @@ from aries_cloudagent.storage.error import StorageNotFoundError
 
 from .util import generate_model_schema, admin_only
 
-PROTOCOL = 'did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/admin-connections/0.1'
+PROTOCOL = (
+    'https://github.com/hyperledger/aries-toolbox/'
+    'tree/master/docs/admin-connections/0.1'
+)
 
 # Message Types
-CONNECTION_GET_LIST = '{}/connection-get-list'.format(PROTOCOL)
-CONNECTION_LIST = '{}/connection-list'.format(PROTOCOL)
-CONNECTION_GET = '{}/connection-get'.format(PROTOCOL)
+GET_LIST = '{}/get-list'.format(PROTOCOL)
+LIST = '{}/list'.format(PROTOCOL)
+UPDATE = '{}/update'.format(PROTOCOL)
 CONNECTION = '{}/connection'.format(PROTOCOL)
-INVITATION_GET_LIST = '{}/invitation-get-list'.format(PROTOCOL)
-INVITATION_LIST = '{}/invitation-list'.format(PROTOCOL)
-CREATE_INVITATION = '{}/create-invitation'.format(PROTOCOL)
-INVITATION = '{}/invitation'.format(PROTOCOL)
+DELETE = '{}/delete'.format(PROTOCOL)
+DELETED = '{}/deleted'.format(PROTOCOL)
 RECEIVE_INVITATION = '{}/receive-invitation'.format(PROTOCOL)
-ACCEPT_INVITATION = '{}/accept-invitation'.format(PROTOCOL)
-ACCEPT_REQUEST = '{}/accept-request'.format(PROTOCOL)
-ESTABLISH_INBOUND = '{}/establish-inbound'.format(PROTOCOL)
-DELETE_CONNECTION = '{}/delete'.format(PROTOCOL)
-UPDATE_CONNECTION = '{}/update'.format(PROTOCOL)
-CONNECTION_ACK = '{}/ack'.format(PROTOCOL)
 
 # Message Type string to Message Class map
 MESSAGE_TYPES = {
-    CONNECTION_GET_LIST:
-        'acapy_plugin_toolbox.connections'
-        '.ConnectionGetList',
-    CONNECTION_LIST:
-        'acapy_plugin_toolbox.connections'
-        '.ConnectionList',
-    CONNECTION_GET:
-        'acapy_plugin_toolbox.connections'
-        '.ConnectionGet',
-    CONNECTION:
-        'acapy_plugin_toolbox.connections'
-        '.Connnection',
-    CREATE_INVITATION:
-        'acapy_plugin_toolbox.connections'
-        '.CreateInvitation',
-    INVITATION_GET_LIST:
-        'acapy_plugin_toolbox.connections'
-        '.InvitationGetList',
-    INVITATION:
-        'acapy_plugin_toolbox.connections'
-        '.Invitation',
-    RECEIVE_INVITATION:
-        'acapy_plugin_toolbox.connections'
-        '.ReceiveInvitation',
-    ACCEPT_INVITATION:
-        'acapy_plugin_toolbox.connections'
-        '.AcceptInvitation',
-    ACCEPT_REQUEST:
-        'acapy_plugin_toolbox.connections'
-        '.AcceptRequest',
-    ESTABLISH_INBOUND:
-        'acapy_plugin_toolbox.connections'
-        '.EstablishInbound',
-    DELETE_CONNECTION:
-        'acapy_plugin_toolbox.connections'
-        '.DeleteConnection',
-    CONNECTION_ACK:
-        'acapy_plugin_toolbox.connections'
-        '.ConnectionAck',
-    UPDATE_CONNECTION:
-        'acapy_plugin_toolbox.connections'
-        '.UpdateConnection',
+    GET_LIST: 'acapy_plugin_toolbox.connections.GetList',
+    LIST: 'acapy_plugin_toolbox.connections.List',
+    UPDATE: 'acapy_plugin_toolbox.connections.Update',
+    CONNECTION: 'acapy_plugin_toolbox.connections.Connnection',
+    DELETE: 'acapy_plugin_toolbox.connections.Delete',
+    DELETED: 'acapy_plugin_toolbox.connections.Deleted',
+    RECEIVE_INVITATION: 'acapy_plugin_toolbox.connections.'
+                        'ReceiveInvitation',
 }
 
 
 async def setup(
         context: InjectionContext,
-        protocol_registry: ProblemReport = None
+        protocol_registry: ProtocolRegistry = None
 ):
     """Setup the connections plugin."""
     if not protocol_registry:
         protocol_registry = await context.inject(ProtocolRegistry)
+
     protocol_registry.register_message_types(
         MESSAGE_TYPES
     )
 
 
-ConnectionGetList, ConnectionGetListSchema = generate_model_schema(
-    name='ConnectionGetList',
-    handler='acapy_plugin_toolbox.connections.ConnectionGetListHandler',
-    msg_type=CONNECTION_GET_LIST,
-    schema={
-        'initiator': fields.Str(
-            validate=validate.OneOf(['self', 'external']),
-            required=False,
-        ),
-        'invitation_key': fields.Str(required=False),
-        'my_did': fields.Str(required=False),
-        'state': fields.Str(
-            validate=validate.OneOf([
-                'init',
-                'invitation',
-                'request',
-                'response',
-                'active',
-                'error',
-                'inactive'
-            ]),
-            required=False
-        ),
-        'their_did': fields.Str(required=False),
-        'their_role': fields.Str(required=False)
-    }
-)
-
-ConnectionList, ConnectionListSchema = generate_model_schema(
-    name='ConnectionList',
-    handler='acapy_plugin_toolbox.util.PassHandler',
-    msg_type=CONNECTION_LIST,
-    schema={
-        'results': fields.List(
-            fields.Nested(ConnectionRecordSchema),
-            required=True
-        )
-    }
-)
-
-ConnectionGet, ConnectionGetSchema = generate_model_schema(
-    name='ConnectionGet',
-    handler='acapy_plugin_toolbox.connections.ConnectionGetHandler',
-    msg_type=CONNECTION_GET,
-    schema={
-        'connection_id': fields.Str(required=True)
-    }
-)
+BaseConnectionSchema = Schema.from_dict({
+    'label': fields.Str(required=True),
+    'my_did': fields.Str(required=True),
+    'connection_id': fields.Str(required=True),
+    'state': fields.Str(
+        validate=validate.OneOf([
+            'pending',
+            'active',
+            'error'
+        ]),
+        required=True
+    ),
+    'their_did': fields.Str(required=False),  # May be missing if pending
+    'role': fields.Str(required=False),
+    'raw_repr': fields.Dict(required=False)
+})
 
 Connection, ConnectionSchema = generate_model_schema(
     name='Connection',
     handler='acapy_plugin_toolbox.util.PassHandler',
     msg_type=CONNECTION,
-    schema={
-        'connection': fields.Nested(ConnectionRecordSchema, required=True),
-    }
+    schema=BaseConnectionSchema
 )
 
-InvitationGetList, InvitationGetListSchema = generate_model_schema(
-    name='InvitationGetList',
-    handler='acapy_plugin_toolbox.connections.InvitationGetListHandler',
-    msg_type=INVITATION_GET_LIST,
+
+def conn_record_to_message_repr(conn: ConnectionRecord) -> Dict[str, Any]:
+    """Map ConnectionRecord onto Connection."""
+    def _state_map(state: str) -> str:
+        if state in ('active', 'response'):
+            return 'active'
+        if state == 'error':
+            return 'error'
+        return 'pending'
+
+    return {
+        'label': conn.their_label,
+        'my_did': conn.my_did,
+        'their_did': conn.their_did,
+        'state': _state_map(conn.state),
+        'role': conn.their_role,
+        'connection_id': conn.connection_id,
+        'raw_repr': conn.serialize()
+    }
+
+
+GetList, GetListSchema = generate_model_schema(
+    name='GetList',
+    handler='acapy_plugin_toolbox.connections.GetListHandler',
+    msg_type=GET_LIST,
     schema={
-        'initiator': fields.Str(
-            validate=validate.OneOf(['self', 'external']),
-            required=False,
-        ),
-        'invitation_key': fields.Str(required=False),
         'my_did': fields.Str(required=False),
-        'their_did': fields.Str(required=False),
-        'their_role': fields.Str(required=False)
-    }
-)
-
-CreateInvitation, CreateInvitationSchema = generate_model_schema(
-    name='CreateInvitation',
-    handler='acapy_plugin_toolbox.connections.CreateInvitationHandler',
-    msg_type=CREATE_INVITATION,
-    schema={
-        'label': fields.Str(required=False),
-        'role': fields.Str(required=False),
-        'accept': fields.Str(
-            required=False,
-            validate=validate.OneOf(['none', 'auto'])
+        'state': fields.Str(
+            validate=validate.OneOf([
+                'pending',
+                'active',
+                'error',
+            ]),
+            required=False
         ),
-        'public': fields.Boolean(missing=False),
-        'multi_use': fields.Boolean(missing=False)
+        'their_did': fields.Str(required=False),
+        'role': fields.Str(required=False)
     }
 )
 
-Invitation, InvitationSchema = generate_model_schema(
-    name='Invitation',
+
+List, ListSchema = generate_model_schema(
+    name='List',
     handler='acapy_plugin_toolbox.util.PassHandler',
-    msg_type=INVITATION,
+    msg_type=LIST,
     schema={
-        'connection_id': fields.Str(required=True),
-        'invitation': fields.Str(required=True),
-        'invitation_url': fields.Str(required=True)
+        'connections': fields.List(
+            fields.Nested(BaseConnectionSchema),
+            required=True
+        )
     }
 )
 
-InvitationList, InvitationListSchema = generate_model_schema(
-    name='InvitationList',
-    handler='acapy_plugin_toolbox.util.PassHandler',
-    msg_type=INVITATION_LIST,
-    schema={
-        'results': fields.List(fields.Dict(
-            connection=fields.Nested(ConnectionSchema),
-            invitation=fields.Nested(InvitationSchema)
+
+class GetListHandler(BaseHandler):
+    """Handler for get connection list request."""
+
+    @admin_only
+    async def handle(self, context: RequestContext, responder: BaseResponder):
+        """Handle get connection list request."""
+
+        tag_filter = dict(
+            filter(lambda item: item[1] is not None, {
+                'my_did': context.message.my_did,
+                'their_did': context.message.their_did,
+            }.items())
+        )
+        post_filter = dict(filter(
+            lambda item: item[1] is not None,
+            {
+                'their_role': context.message.role
+            }.items()
         ))
-    }
-)
-
-class CreateInvitationHandler(BaseHandler):
-    """Handler for create invitation request."""
-
-    @admin_only
-    async def handle(self, context: RequestContext, responder: BaseResponder):
-        """Handle create invitation request."""
-        connection_mgr = ConnectionManager(context)
-        connection, invitation = await connection_mgr.create_invitation(
-            my_label=context.message.label,
-            their_role=context.message.role,
-            accept=context.message.accept,
-            multi_use=bool(context.message.multi_use),
-            public=bool(context.message.public),
+        # TODO: Filter on state (needs mapping back to ACA-Py connection states)
+        records = await ConnectionRecord.query(
+            context, tag_filter, post_filter
         )
-        invite_response = Invitation(
-            connection_id=connection and connection.connection_id,
-            invitation=invitation.serialize(),
-            invitation_url=invitation.to_url(),
-        )
-        invite_response.assign_thread_from(context.message)
-        await responder.send_reply(invite_response)
+        results = [
+            Connection(**conn_record_to_message_repr(record))
+            for record in records
+        ]
+        connection_list = List(connections=results)
+        connection_list.assign_thread_from(context.message)
+        await responder.send_reply(connection_list)
 
 
-ReceiveInvitation, ReceiveInvitationSchema = generate_model_schema(
-    name='ReceiveInvitation',
-    handler='acapy_plugin_toolbox.connections.ReceiveInvitationHandler',
-    msg_type=RECEIVE_INVITATION,
-    schema={
-        'invitation': fields.Str(required=True),
-        'accept': fields.Str(
-            validate=validate.OneOf(['none', 'auto']),
-            missing=False
-        )
-    }
-)
-
-
-class ReceiveInvitationHandler(BaseHandler):
-    """Handler for receive invitation request."""
-
-    @admin_only
-    async def handle(self, context: RequestContext, responder: BaseResponder):
-        """Handle recieve invitation request."""
-        connection_mgr = ConnectionManager(context)
-        invitation = ConnectionInvitation.from_url(context.message.invitation)
-        connection = await connection_mgr.receive_invitation(
-            invitation, accept=context.message.accept
-        )
-        connection_resp = Connection(connection=connection)
-        await responder.send_reply(connection_resp)
-
-
-AcceptInvitation, AcceptInvitationSchema = generate_model_schema(
-    name='AcceptInvitation',
-    handler='acapy_plugin_toolbox.connections.AcceptInvitationHandler',
-    msg_type=ACCEPT_INVITATION,
-    schema={
-        'connection_id': fields.Str(required=True),
-        'my_endpoint': fields.Str(required=False),
-        'my_label': fields.Str(required=False),
-    }
-)
-
-AcceptRequest, AcceptRequestSchema = generate_model_schema(
-    name='AcceptRequest',
-    handler='acapy_plugin_toolbox.connections.AcceptRequestHandler',
-    msg_type=ACCEPT_REQUEST,
-    schema={
-        'connection_id': fields.Str(required=True),
-        'my_endpoint': fields.Str(required=False),
-    }
-)
-
-EstablishInbound, EstablishInboundSchema = generate_model_schema(
-    name='EstablishInbound',
-    handler='acapy_plugin_toolbox.connections.EstablishInboundHandler',
-    msg_type=ESTABLISH_INBOUND,
-    schema={
-        'connection_id': fields.Str(required=True),
-        'ref_id': fields.Str(required=True),
-    }
-)
-
-DeleteConnection, DeleteConnectionSchema = generate_model_schema(
-    name='DeleteConnection',
-    handler='acapy_plugin_toolbox.connections.DeleteConnectionHandler',
-    msg_type=DELETE_CONNECTION,
-    schema={
-        'connection_id': fields.Str(required=True),
-    }
-)
-
-ConnectionAck, ConnectionAckSchema = generate_model_schema(
-    name='ConnectionAck',
-    handler='acapy_plugin_toolbox.util.PassHandler',
-    msg_type=CONNECTION_ACK,
-    schema={}
-)
-
-UpdateConnection, UpdateConnectionSchema = generate_model_schema(
-    name='UpdateConnection',
-    handler='acapy_plugin_toolbox.connections.UpdateConnectionHandler',
-    msg_type=UPDATE_CONNECTION,
+Update, UpdateSchema = generate_model_schema(
+    name='Update',
+    handler='acapy_plugin_toolbox.connections.UpdateHandler',
+    msg_type=UPDATE,
     schema={
         'connection_id': fields.Str(required=True),
         'label': fields.Str(required=False),
@@ -322,7 +184,57 @@ UpdateConnection, UpdateConnectionSchema = generate_model_schema(
 )
 
 
-class DeleteConnectionHandler(BaseHandler):
+class UpdateHandler(BaseHandler):
+    """Handler for update connection request."""
+
+    @admin_only
+    async def handle(self, context: RequestContext, responder: BaseResponder):
+        """Handle update connection request."""
+        try:
+            connection = await ConnectionRecord.retrieve_by_id(
+                context,
+                context.message.connection_id
+            )
+        except StorageNotFoundError:
+            report = ProblemReport(
+                explain_ltxt='Connection not found.',
+                who_retries='none'
+            )
+            report.assign_thread_from(context.message)
+            await responder.send_reply(report)
+
+        new_label = context.message.label or connection.their_label
+        connection.their_label = new_label
+        new_role = context.message.role or connection.their_role
+        connection.their_role = new_role
+        await connection.save(context, reason="Update request received.")
+        conn_response = Connection(
+            **conn_record_to_message_repr(connection)
+        )
+        conn_response.assign_thread_from(context.message)
+        await responder.send_reply(conn_response)
+
+
+Delete, DeleteSchema = generate_model_schema(
+    name='Delete',
+    handler='acapy_plugin_toolbox.connections.DeleteHandler',
+    msg_type=DELETE,
+    schema={
+        'connection_id': fields.Str(required=True),
+    }
+)
+
+Deleted, DeletedSchema = generate_model_schema(
+    name='Deleted',
+    handler='acapy_plugin_toolbox.util.PassHandler',
+    msg_type=DELETED,
+    schema={
+        'connection_id': fields.Str(required=True),
+    }
+)
+
+
+class DeleteHandler(BaseHandler):
     """Handler for delete connection request."""
 
     @admin_only
@@ -354,118 +266,35 @@ class DeleteConnectionHandler(BaseHandler):
             return
 
         await connection.delete_record(context)
-        ack = ConnectionAck()
-        ack.assign_thread_from(context.message)
-        await responder.send_reply(ack)
+        deleted = Deleted(connection_id=connection.connection_id)
+        deleted.assign_thread_from(context.message)
+        await responder.send_reply(deleted)
 
 
-class UpdateConnectionHandler(BaseHandler):
-    """Handler for update connection request."""
-
-    @admin_only
-    async def handle(self, context: RequestContext, responder: BaseResponder):
-        """Handle update connection request."""
-        try:
-            connection = await ConnectionRecord.retrieve_by_id(
-                context,
-                context.message.connection_id
-            )
-        except StorageNotFoundError:
-            report = ProblemReport(
-                explain_ltxt='Connection not found.',
-                who_retries='none'
-            )
-            report.assign_thread_from(context.message)
-            await responder.send_reply(report)
-
-        new_label = context.message.label or connection.their_label
-        connection.their_label = new_label
-        new_role = context.message.role or connection.their_role
-        connection.their_role = new_role
-        await connection.save(context, reason="Update request received.")
-        conn_response = Connection(connection=connection)
-        conn_response.assign_thread_from(context.message)
-        await responder.send_reply(conn_response)
-
-
-class ConnectionGetListHandler(BaseHandler):
-    """Handler for get connection list request."""
-
-    @admin_only
-    async def handle(self, context: RequestContext, responder: BaseResponder):
-        """Handle get connection list request."""
-
-        def connection_sort_key(conn):
-            """Get the sorting key for a particular connection."""
-            if conn["state"] == ConnectionRecord.STATE_INACTIVE:
-                pfx = "2"
-            elif conn["state"] == ConnectionRecord.STATE_INVITATION:
-                pfx = "1"
-            else:
-                pfx = "0"
-            return pfx + conn["created_at"]
-
-        tag_filter = dict(
-            filter(lambda item: item[1] is not None, {
-                'my_did': context.message.my_did,
-                'their_did': context.message.their_did,
-            }.items())
+ReceiveInvitation, ReceiveInvitationSchema = generate_model_schema(
+    name='ReceiveInvitation',
+    handler='acapy_plugin_toolbox.connections.ReceiveInvitationHandler',
+    msg_type=RECEIVE_INVITATION,
+    schema={
+        'invitation': fields.Str(required=True),
+        'auto_accept': fields.Bool(
+            missing=False
         )
-        post_filter = dict(filter(
-            lambda item: item[1] is not None,
-            {
-                'initiator': context.message.initiator,
-                'state': context.message.state,
-                'their_role': context.message.their_role
-            }.items()
-        ))
-        records = await ConnectionRecord.query(context, tag_filter, post_filter)
-        results = [record.serialize() for record in records]
-        results.sort(key=connection_sort_key)
-        connection_list = ConnectionList(results=results)
-        connection_list.assign_thread_from(context.message)
-        await responder.send_reply(connection_list)
+    }
+)
 
 
-class InvitationGetListHandler(BaseHandler):
-    """Handler for get invitation list request."""
+class ReceiveInvitationHandler(BaseHandler):
+    """Handler for receive invitation request."""
 
     @admin_only
     async def handle(self, context: RequestContext, responder: BaseResponder):
-        """Handle get invitation list request."""
-
-        tag_filter = dict(
-            filter(lambda item: item[1] is not None, {
-                'my_did': context.message.my_did,
-                'their_did': context.message.their_did,
-            }.items())
+        """Handle recieve invitation request."""
+        connection_mgr = ConnectionManager(context)
+        invitation = ConnectionInvitation.from_url(context.message.invitation)
+        connection = await connection_mgr.receive_invitation(
+            invitation,
+            accept=('auto' if context.message.auto_accept else 'none')
         )
-        post_filter = dict(filter(
-            lambda item: item[1] is not None,
-            {
-                'initiator': context.message.initiator,
-                'state': 'invitation',
-                'their_role': context.message.their_role
-            }.items()
-        ))
-        records = await ConnectionRecord.query(context, tag_filter, post_filter)
-        results = []
-        for connection in records:
-            try:
-                invitation = await connection.retrieve_invitation(context)
-            except StorageNotFoundError:
-                continue
-
-            row = {
-                'connection': connection.serialize(),
-                'invitation': {
-                    'connection_id': connection and connection.connection_id,
-                    'invitation': invitation.serialize(),
-                    'invitation_url': invitation.to_url(),
-                }
-            }
-            results.append(row)
-
-        invitation_list = InvitationList(results=results)
-        invitation_list.assign_thread_from(context.message)
-        await responder.send_reply(invitation_list)
+        connection_resp = Connection(**conn_record_to_message_repr(connection))
+        await responder.send_reply(connection_resp)
