+++ acapy_plugin_toolbox/group/all.py
@@ -4,7 +4,6 @@ from aries_cloudagent.config.injection_context import InjectionContext
 from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 
 from ..connections import setup as connection_setup
-from ..connections_new import setup as connection_new_setup
 from ..credential_definitions import setup as cred_def_setup
 from ..schemas import setup as schema_setup
 from ..dids import setup as did_setup
@@ -20,7 +19,6 @@ async def setup(context: InjectionContext):
     """Setup Toolbox Plugin."""
     protocol_registry = await context.inject(ProtocolRegistry)
     await connection_setup(context, protocol_registry)
-    await connection_new_setup(context, protocol_registry)
     await cred_def_setup(context, protocol_registry)
     await schema_setup(context, protocol_registry)
     await did_setup(context, protocol_registry)
